Modify import statement to not abbrieviate module name
for the sake of 2 characters it is not worth the difficulty.
Caused by trying to get eval(repr(obj)) to work which would call:
eval(image_info.ImageInfo(...)), however, the module was imported as img_info and hence could obviously not be found.

Furthermore, there was an error with the repr in that it was trying to call ImageObject instead of ImageInfo