Change how PIVImage objects are initialised
Changed the initialisation of the object to require the arrays IA, IB, and possibly mask to be passed in, instead of the flowtype number and image number.

Changing in this way will allow for easier testing of the object since arrays with known properties can be mocked far easier. This will require a new 'factory' function to create image object in a similar manner to before. Something like:
obj = initialisePIVImageFromFlowtype(im_details, im_number)

Also changed is the class name from piv_image to PIVImage to follow better naming conventions.