Working piv_image.get_region(x, y, r) method
This method gets the pixel intensity and mask values in the region
(x-rad:x+rad, y-rad:y+rad) and returns the three arrays.
if x/y +- rad extends beyond the image then this region is padded with 0's

The binary flag 'has_mask' is also stored in the object to facilitate checking
this value before trying to extract mask values. If there is no mask,
then all these values will be 0 anyway so we can just create zeros(WS)