Update densepredictor to apply mask more rigorously
applies the mask on initialisation and when mathematical operators are applied.

Mask behaviour for mathematical operators:
The left hand mask is always taken to be the new mask,
i.e. c = a + b
c.mask == a.mask which does not necessarily equal b.mask
If the masks are not the same, then a warning is issued, but the lhs mask taken
Consequence:
a has an arbitrary mask
b has no mask
c = a +-*/ b -> all areas with a mask in a are still 0 in c
c = b +-*/ a -> all areas with a mask in a are == b in