Change operator behaviour
If trying to add/subtract/multiply/divide two different densepredictors
then the two must have identical masks.

The previous approach, taking the lhs mask from a + b, has a lot of decisions
to be made, particularly for division and multiplication.
e.g. if a mask region adds with a non-mask region, should this be mask
or should if be equal to the value of the non-mask region?
Either could be equally valid, however, this then requires making decisions
now about how this should be enforced.
Since in the vast majority of cases (all that I can currently think of)
we will always be adding objects based on the same image, and hence mask,
I have decided to just raise an error if they are different.

In addition, I have refactored some of the test code:
u = np.arange(1, 81)
u = np.reshape(u, (9,9))
===
u = np.arange(1, 81).reshape((9,9))