Method to return all of a specified 'value' from Distribution
Distribution.values(property) returns a list of all
self.windows.property

uses the __dict__ to access a specific property as this is quicker
than something like getattr, however, may cause some issues with
inheritance. Using this method for now as it doesn't seem to cause any
issues, however, this should be noted in case of future issues.