New utilities file
contains utility, generic useful functions.

First two functions include an elementwise_diff to find i.e.
A[:-1] - A[1:]     (A[i+1] - A[i], for i = 1 ... N-1)

The second utility function reshapes a flat array of unknown dimensions
into a 2d array, by looking at the spacing in x and y.
Checks if the array 'makes sense' and throws a ValueError if it doesn't