Bug fix and test for vector validation
In the previous testing of validation, the u and v going in were made
from numpy arrays. When testing as part of the object, because
'distribution.get_values' returns a list, the indexing of the form
u_nb, v_nb = u[row[1:]], v[row[1:]]
was not working because, i think, the index needs to be a pure int,
whereas here row[1:] is an ndarray of type int64.
Converting all values out of get_values to numpy array solves this
since we are dealing purely with numpy arrays.