Update seeding density calculations
For clarity, calc_seeding_density now takes the original image, detects
any particles and calculates the seeding density over the domain. This
I think is more in line with what is expected of the function.

In doing this, a function calculate_local_mean_value is created which
is effectively a unity weighted 2D convolution, which considers the
presence of a mask.