Seeding density estimation (#54)
* Start modifying particle generation to make it easier to vary density

* Separate particle generation code from particle displacement

Allows us to:
1. generate individual particle images more efficiently for testing
e.g. particle detection
2. more easily generate non-uniform seeding distributions

* Update docstring and remove 'main' script

* Method to generate particle images with a varying seeding density

* Consider masks when detecting particles with max filter

Make sure masks are considered when detecting particle images using the
thresholded maximum filter.

Ensure that the threshold calculation considers the mask as well as
that there shouldn't be any particles in the masked area

* Update seeding density calculations

For clarity, calc_seeding_density now takes the original image, detects
any particles and calculates the seeding density over the domain. This
I think is more in line with what is expected of the function.

In doing this, a function calculate_local_mean_value is created which
is effectively a unity weighted 2D convolution, which considers the
presence of a mask.

* Add dispatcher for detect particles