Bugfix: create CorrWindow list from grid
Update such that the cw list can be created from a two dimensional grid
as well.

The bug that was raising is that by adding WS.ravel(), if WS was a list
i.e. a scalar was passed and isinstance(WS, int) was triggered, then
there is no function ravel for a list.