Adaptive inc stippling (#61)
* Add Disk class

* Method is running

* Initial setup for cythonising

Simply placed everything into a cython file, nothing typed

original comparison
~2s for python
~1.8s for cython

linear ramp density, 1000 points

* Cythonised SummedAreaTable

~2s python
~1.75 cython

* Basic cythonising of everything else

This is done in one go because, when you cdef a function, it appears
only other cdef'd functions can actually use it - even if they're in
cython already.

I think it is to do with whether a variable has been typed or not.

If a variable in a non-cdef'd function is not typed, it acts as a
python variable. Therefore, it cannot access the cdef function.
If it is typed, then the cython code knows it can access the cdef func

* Various optimisations

~2s python
~0.1s Cython