Bugfix: Don't raise error when trying to split an already split cell
It is possible that this can happen in a valid series of events.

Consider a list of leaf cells is created from a mesh with a significant
amount of tier variability.
Going through this list and splitting a higher tier cell may cascade
into splitting lower tier cells.
The original list may contain some of these lower tier cells and so
when it comes to splitting this an error is incorrectly raised.

A (perhaps) more robust approach would be that when splitting a higher
tier cell cascades into lower tier cells, that the lower tier cells are
removed from the current list to be split - though this would be quite
complex to implement.